requirements.txt
streamlit
yfinance
pandas
numpy
scikit-learn
tensorflow
matplotlib
Pillow
streamlit-option-menu